requirements.txt
chainlit>=0.7.700
google-cloud-discoveryengine>=0.11.10
python-dotenv>=1.0.0
